+++ tgbotapi.behaviour_builder/src/commonMain/kotlin/dev/inmo/tgbotapi/extensions/behaviour_builder/triggers_handling/CommandHandling.kt
@@ -5,6 +5,8 @@ package dev.inmo.tgbotapi.extensions.behaviour_builder.triggers_handling
 import dev.inmo.tgbotapi.extensions.behaviour_builder.*
 import dev.inmo.tgbotapi.extensions.behaviour_builder.filters.CommonMessageFilterExcludeMediaGroups
 import dev.inmo.tgbotapi.extensions.behaviour_builder.filters.MessageFilterByChat
+import dev.inmo.tgbotapi.extensions.behaviour_builder.utils.botInfo
+import dev.inmo.tgbotapi.extensions.behaviour_builder.utils.containsCommand
 import dev.inmo.tgbotapi.extensions.behaviour_builder.utils.marker_factories.ByChatMessageMarkerFactory
 import dev.inmo.tgbotapi.extensions.behaviour_builder.utils.marker_factories.MarkerFactory
 import dev.inmo.tgbotapi.extensions.behaviour_builder.utils.times
@@ -15,6 +17,7 @@ import dev.inmo.tgbotapi.extensions.utils.extensions.parseCommandsWithNamedArgs
 import dev.inmo.tgbotapi.types.BotCommand
 import dev.inmo.tgbotapi.types.message.content.TextContent
 import dev.inmo.tgbotapi.types.message.content.TextMessage
+import dev.inmo.tgbotapi.types.message.textsources.BotCommandTextSource
 import dev.inmo.tgbotapi.types.update.abstracts.Update
 import dev.inmo.tgbotapi.utils.launchWithBotLogger
 import kotlinx.coroutines.Job
@@ -27,6 +30,7 @@ internal fun <BC : BehaviourContext> BC.commandUncounted(
     subcontextUpdatesFilter: CustomBehaviourContextAndTwoTypesReceiver<BC, Boolean, TextMessage, Update>? = MessageFilterByChat,
     markerFactory: MarkerFactory<in TextMessage, Any>? = ByChatMessageMarkerFactory,
     additionalSubcontextInitialAction: CustomBehaviourContextAndTwoTypesReceiver<BC, Unit, Update, TextMessage>? = null,
+    excludeCommandsToOtherBots: Boolean = true,
     scenarioReceiver: CustomBehaviourContextAndTypeReceiver<BC, Unit, TextMessage>
 ): Job = onText(
     CommonMessageFilter<TextContent> { message ->
@@ -39,6 +43,16 @@ internal fun <BC : BehaviourContext> BC.commandUncounted(
         }
         sizeRequirement && textSources.any {
             commandRegex.matches(it.botCommandTextSourceOrNull() ?.command ?: return@any false)
+        }
+    }.let {
+        if (excludeCommandsToOtherBots) {
+            it * lambda@{
+                with(it.content.textSources) {
+                    containsCommand(commandRegex)
+                }
+            }
+        } else {
+            it
         }
     }.let {
         initialFilter ?.times(it) ?: it
@@ -56,17 +70,19 @@ fun <BC : BehaviourContext> BC.command(
     subcontextUpdatesFilter: CustomBehaviourContextAndTwoTypesReceiver<BC, Boolean, TextMessage, Update>? = MessageFilterByChat,
     markerFactory: MarkerFactory<in TextMessage, Any>? = ByChatMessageMarkerFactory,
     additionalSubcontextInitialAction: CustomBehaviourContextAndTwoTypesReceiver<BC, Unit, Update, TextMessage>? = null,
+    excludeCommandsToOtherBots: Boolean = true,
     scenarioReceiver: CustomBehaviourContextAndTypeReceiver<BC, Unit, TextMessage>
 ): Job = launchInNewSubContext {
     runCatching {
        this@launchInNewSubContext.commandUncounted(
-            commandRegex,
-            requireOnlyCommandInMessage,
-            initialFilter,
-            subcontextUpdatesFilter,
-            markerFactory,
-            additionalSubcontextInitialAction,
-            scenarioReceiver
+           commandRegex = commandRegex,
+           requireOnlyCommandInMessage = requireOnlyCommandInMessage,
+           initialFilter = initialFilter,
+           subcontextUpdatesFilter = subcontextUpdatesFilter,
+           markerFactory = markerFactory,
+           additionalSubcontextInitialAction = additionalSubcontextInitialAction,
+           excludeCommandsToOtherBots = excludeCommandsToOtherBots,
+           scenarioReceiver = scenarioReceiver
         )
     }.onFailure {
         triggersHolder.handleableCommandsHolder.unregisterHandleable(commandRegex)
@@ -94,8 +110,18 @@ fun <BC : BehaviourContext> BC.command(
     subcontextUpdatesFilter: CustomBehaviourContextAndTwoTypesReceiver<BC, Boolean, TextMessage, Update>? = MessageFilterByChat,
     markerFactory: MarkerFactory<in TextMessage, Any>? = ByChatMessageMarkerFactory,
     additionalSubcontextInitialAction: CustomBehaviourContextAndTwoTypesReceiver<BC, Unit, Update, TextMessage>? = null,
+    excludeCommandsToOtherBots: Boolean = true,
     scenarioReceiver: CustomBehaviourContextAndTypeReceiver<BC, Unit, TextMessage>
-) = command(command.toRegex(), requireOnlyCommandInMessage, initialFilter, subcontextUpdatesFilter, markerFactory, additionalSubcontextInitialAction, scenarioReceiver)
+) = command(
+    commandRegex = command.toRegex(),
+    requireOnlyCommandInMessage = requireOnlyCommandInMessage,
+    initialFilter = initialFilter,
+    subcontextUpdatesFilter = subcontextUpdatesFilter,
+    markerFactory = markerFactory,
+    additionalSubcontextInitialAction = additionalSubcontextInitialAction,
+    excludeCommandsToOtherBots = excludeCommandsToOtherBots,
+    scenarioReceiver = scenarioReceiver
+)
 
 /**
  * @param [markerFactory] **Pass null to handle requests fully parallel**. Will be used to identify different "stream".
@@ -109,8 +135,18 @@ fun <BC : BehaviourContext> BC.command(
     subcontextUpdatesFilter: CustomBehaviourContextAndTwoTypesReceiver<BC, Boolean, TextMessage, Update>? = MessageFilterByChat,
     markerFactory: MarkerFactory<in TextMessage, Any>? = ByChatMessageMarkerFactory,
     additionalSubcontextInitialAction: CustomBehaviourContextAndTwoTypesReceiver<BC, Unit, Update, TextMessage>? = null,
+    excludeCommandsToOtherBots: Boolean = true,
     scenarioReceiver: CustomBehaviourContextAndTypeReceiver<BC, Unit, TextMessage>
-) = command(botCommand.command, requireOnlyCommandInMessage, initialFilter, subcontextUpdatesFilter, markerFactory, additionalSubcontextInitialAction, scenarioReceiver)
+) = command(
+    command = botCommand.command,
+    requireOnlyCommandInMessage = requireOnlyCommandInMessage,
+    initialFilter = initialFilter,
+    subcontextUpdatesFilter = subcontextUpdatesFilter,
+    markerFactory = markerFactory,
+    additionalSubcontextInitialAction = additionalSubcontextInitialAction,
+    excludeCommandsToOtherBots = excludeCommandsToOtherBots,
+    scenarioReceiver = scenarioReceiver
+)
 
 /**
  * @param [markerFactory] **Pass null to handle requests fully parallel**. Will be used to identify different "stream".
@@ -124,8 +160,18 @@ fun <BC : BehaviourContext> BC.onCommand(
     subcontextUpdatesFilter: CustomBehaviourContextAndTwoTypesReceiver<BC, Boolean, TextMessage, Update>? = MessageFilterByChat,
     markerFactory: MarkerFactory<in TextMessage, Any>? = ByChatMessageMarkerFactory,
     additionalSubcontextInitialAction: CustomBehaviourContextAndTwoTypesReceiver<BC, Unit, Update, TextMessage>? = null,
+    excludeCommandsToOtherBots: Boolean = true,
     scenarioReceiver: CustomBehaviourContextAndTypeReceiver<BC, Unit, TextMessage>
-): Job = command(commandRegex, requireOnlyCommandInMessage, initialFilter, subcontextUpdatesFilter, markerFactory, additionalSubcontextInitialAction, scenarioReceiver)
+): Job = command(
+    commandRegex = commandRegex,
+    requireOnlyCommandInMessage = requireOnlyCommandInMessage,
+    initialFilter = initialFilter,
+    subcontextUpdatesFilter = subcontextUpdatesFilter,
+    markerFactory = markerFactory,
+    additionalSubcontextInitialAction = additionalSubcontextInitialAction,
+    excludeCommandsToOtherBots = excludeCommandsToOtherBots,
+    scenarioReceiver = scenarioReceiver
+)
 
 /**
  * @param [markerFactory] **Pass null to handle requests fully parallel**. Will be used to identify different "stream".
@@ -139,8 +185,18 @@ fun <BC : BehaviourContext> BC.onCommand(
     subcontextUpdatesFilter: CustomBehaviourContextAndTwoTypesReceiver<BC, Boolean, TextMessage, Update>? = MessageFilterByChat,
     markerFactory: MarkerFactory<in TextMessage, Any>? = ByChatMessageMarkerFactory,
     additionalSubcontextInitialAction: CustomBehaviourContextAndTwoTypesReceiver<BC, Unit, Update, TextMessage>? = null,
+    excludeCommandsToOtherBots: Boolean = true,
     scenarioReceiver: CustomBehaviourContextAndTypeReceiver<BC, Unit, TextMessage>
-): Job = onCommand(command.toRegex(), requireOnlyCommandInMessage, initialFilter, subcontextUpdatesFilter, markerFactory, additionalSubcontextInitialAction, scenarioReceiver)
+): Job = onCommand(
+    commandRegex = command.toRegex(),
+    requireOnlyCommandInMessage = requireOnlyCommandInMessage,
+    initialFilter = initialFilter,
+    subcontextUpdatesFilter = subcontextUpdatesFilter,
+    markerFactory = markerFactory,
+    additionalSubcontextInitialAction = additionalSubcontextInitialAction,
+    excludeCommandsToOtherBots = excludeCommandsToOtherBots,
+    scenarioReceiver = scenarioReceiver
+)
 
 /**
  * @param [markerFactory] **Pass null to handle requests fully parallel**. Will be used to identify different "stream".
@@ -154,8 +210,18 @@ fun <BC : BehaviourContext> BC.onCommand(
     subcontextUpdatesFilter: CustomBehaviourContextAndTwoTypesReceiver<BC, Boolean, TextMessage, Update>? = MessageFilterByChat,
     markerFactory: MarkerFactory<in TextMessage, Any>? = ByChatMessageMarkerFactory,
     additionalSubcontextInitialAction: CustomBehaviourContextAndTwoTypesReceiver<BC, Unit, Update, TextMessage>? = null,
+    excludeCommandsToOtherBots: Boolean = true,
     scenarioReceiver: CustomBehaviourContextAndTypeReceiver<BC, Unit, TextMessage>
-): Job = onCommand(botCommand.command, requireOnlyCommandInMessage, initialFilter, subcontextUpdatesFilter, markerFactory, additionalSubcontextInitialAction, scenarioReceiver)
+): Job = onCommand(
+    command = botCommand.command,
+    requireOnlyCommandInMessage = requireOnlyCommandInMessage,
+    initialFilter = initialFilter,
+    subcontextUpdatesFilter = subcontextUpdatesFilter,
+    markerFactory = markerFactory,
+    additionalSubcontextInitialAction = additionalSubcontextInitialAction,
+    excludeCommandsToOtherBots = excludeCommandsToOtherBots,
+    scenarioReceiver = scenarioReceiver
+)
 
 /**
  * @param [markerFactory] **Pass null to handle requests fully parallel**. Will be used to identify different "stream".
@@ -169,13 +235,15 @@ fun <BC : BehaviourContext> BC.commandWithArgs(
     markerFactory: MarkerFactory<in TextMessage, Any>? = ByChatMessageMarkerFactory,
     additionalSubcontextInitialAction: CustomBehaviourContextAndTwoTypesReceiver<BC, Unit, Update, TextMessage>? = null,
     argsSeparator: Regex = TelegramBotCommandsDefaults.defaultArgsSeparatorRegex,
+    excludeCommandsToOtherBots: Boolean = true,
     scenarioReceiver: CustomBehaviourContextAndTwoTypesReceiver<BC, Unit, TextMessage, Array<String>>
 ) = command(
     commandRegex,
     requireOnlyCommandInMessage = false,
     initialFilter = initialFilter,
     subcontextUpdatesFilter = subcontextUpdatesFilter,
-    markerFactory = markerFactory
+    markerFactory = markerFactory,
+    excludeCommandsToOtherBots = excludeCommandsToOtherBots,
 ) {
     val args = it.parseCommandsWithArgs(argsSeparator = argsSeparator).let { commandsWithArgs ->
         val key = commandsWithArgs.keys.firstOrNull { it.matches(commandRegex) } ?: return@let null
@@ -196,6 +264,7 @@ fun <BC : BehaviourContext> BC.commandWithArgs(
     markerFactory: MarkerFactory<in TextMessage, Any>? = ByChatMessageMarkerFactory,
     additionalSubcontextInitialAction: CustomBehaviourContextAndTwoTypesReceiver<BC, Unit, Update, TextMessage>? = null,
     argsSeparator: Regex = TelegramBotCommandsDefaults.defaultArgsSeparatorRegex,
+    excludeCommandsToOtherBots: Boolean = true,
     scenarioReceiver: CustomBehaviourContextAndTwoTypesReceiver<BC, Unit, TextMessage, Array<String>>
 ) = commandWithArgs(
     command.toRegex(),
@@ -203,6 +272,7 @@ fun <BC : BehaviourContext> BC.commandWithArgs(
     subcontextUpdatesFilter = subcontextUpdatesFilter,
     markerFactory = markerFactory,
     argsSeparator = argsSeparator,
+    excludeCommandsToOtherBots = excludeCommandsToOtherBots,
     scenarioReceiver = scenarioReceiver
 )
 
@@ -218,6 +288,7 @@ fun <BC : BehaviourContext> BC.commandWithArgs(
     markerFactory: MarkerFactory<in TextMessage, Any>? = ByChatMessageMarkerFactory,
     additionalSubcontextInitialAction: CustomBehaviourContextAndTwoTypesReceiver<BC, Unit, Update, TextMessage>? = null,
     argsSeparator: Regex = TelegramBotCommandsDefaults.defaultArgsSeparatorRegex,
+    excludeCommandsToOtherBots: Boolean = true,
     scenarioReceiver: CustomBehaviourContextAndTwoTypesReceiver<BC, Unit, TextMessage, Array<String>>
 ) = commandWithArgs(
     botCommand.command,
@@ -225,6 +296,7 @@ fun <BC : BehaviourContext> BC.commandWithArgs(
     subcontextUpdatesFilter = subcontextUpdatesFilter,
     markerFactory = markerFactory,
     argsSeparator = argsSeparator,
+    excludeCommandsToOtherBots = excludeCommandsToOtherBots,
     scenarioReceiver = scenarioReceiver
 )
 
@@ -241,12 +313,14 @@ fun <BC : BehaviourContext> BC.commandWithNamedArgs(
     additionalSubcontextInitialAction: CustomBehaviourContextAndTwoTypesReceiver<BC, Unit, Update, TextMessage>? = null,
     argsSeparator: Regex = TelegramBotCommandsDefaults.defaultArgsSeparatorRegex,
     nameArgSeparator: Regex = TelegramBotCommandsDefaults.defaultNamesArgsSeparatorRegex,
+    excludeCommandsToOtherBots: Boolean = true,
     scenarioReceiver: CustomBehaviourContextAndTwoTypesReceiver<BC, Unit, TextMessage, List<Pair<String, String>>>
 ) = command(
     commandRegex,
     requireOnlyCommandInMessage = false,
     initialFilter = initialFilter,
     subcontextUpdatesFilter = subcontextUpdatesFilter,
+    excludeCommandsToOtherBots = excludeCommandsToOtherBots,
     markerFactory = markerFactory
 ) {
     val args = it.parseCommandsWithNamedArgs(argsSeparator = argsSeparator, nameArgSeparator = nameArgSeparator).let { commandsWithArgs ->
@@ -269,6 +343,7 @@ fun <BC : BehaviourContext> BC.commandWithNamedArgs(
     additionalSubcontextInitialAction: CustomBehaviourContextAndTwoTypesReceiver<BC, Unit, Update, TextMessage>? = null,
     argsSeparator: Regex = TelegramBotCommandsDefaults.defaultArgsSeparatorRegex,
     nameArgSeparator: Regex = TelegramBotCommandsDefaults.defaultNamesArgsSeparatorRegex,
+    excludeCommandsToOtherBots: Boolean = true,
     scenarioReceiver: CustomBehaviourContextAndTwoTypesReceiver<BC, Unit, TextMessage, List<Pair<String, String>>>
 ) = commandWithNamedArgs(
     command.toRegex(),
@@ -277,6 +352,7 @@ fun <BC : BehaviourContext> BC.commandWithNamedArgs(
     markerFactory = markerFactory,
     argsSeparator = argsSeparator,
     nameArgSeparator = nameArgSeparator,
+    excludeCommandsToOtherBots = excludeCommandsToOtherBots,
     scenarioReceiver = scenarioReceiver
 )
 
@@ -293,6 +369,7 @@ fun <BC : BehaviourContext> BC.commandWithNamedArgs(
     additionalSubcontextInitialAction: CustomBehaviourContextAndTwoTypesReceiver<BC, Unit, Update, TextMessage>? = null,
     argsSeparator: Regex = TelegramBotCommandsDefaults.defaultArgsSeparatorRegex,
     nameArgSeparator: Regex = TelegramBotCommandsDefaults.defaultNamesArgsSeparatorRegex,
+    excludeCommandsToOtherBots: Boolean = true,
     scenarioReceiver: CustomBehaviourContextAndTwoTypesReceiver<BC, Unit, TextMessage, List<Pair<String, String>>>
 ) = commandWithNamedArgs(
     botCommand.command,
@@ -301,6 +378,7 @@ fun <BC : BehaviourContext> BC.commandWithNamedArgs(
     markerFactory = markerFactory,
     argsSeparator = argsSeparator,
     nameArgSeparator = nameArgSeparator,
+    excludeCommandsToOtherBots = excludeCommandsToOtherBots,
     scenarioReceiver = scenarioReceiver
 )
 
@@ -316,6 +394,7 @@ fun <BC : BehaviourContext> BC.onCommandWithArgs(
     markerFactory: MarkerFactory<in TextMessage, Any>? = ByChatMessageMarkerFactory,
     additionalSubcontextInitialAction: CustomBehaviourContextAndTwoTypesReceiver<BC, Unit, Update, TextMessage>? = null,
     argsSeparator: Regex = TelegramBotCommandsDefaults.defaultArgsSeparatorRegex,
+    excludeCommandsToOtherBots: Boolean = true,
     scenarioReceiver: CustomBehaviourContextAndTwoTypesReceiver<BC, Unit, TextMessage, Array<String>>
 ): Job = commandWithArgs(
     commandRegex = commandRegex,
@@ -323,6 +402,7 @@ fun <BC : BehaviourContext> BC.onCommandWithArgs(
     subcontextUpdatesFilter = subcontextUpdatesFilter,
     markerFactory = markerFactory,
     argsSeparator = argsSeparator,
+    excludeCommandsToOtherBots = excludeCommandsToOtherBots,
     scenarioReceiver = scenarioReceiver
 )
 
@@ -338,6 +418,7 @@ fun <BC : BehaviourContext> BC.onCommandWithArgs(
     markerFactory: MarkerFactory<in TextMessage, Any>? = ByChatMessageMarkerFactory,
     additionalSubcontextInitialAction: CustomBehaviourContextAndTwoTypesReceiver<BC, Unit, Update, TextMessage>? = null,
     argsSeparator: Regex = TelegramBotCommandsDefaults.defaultArgsSeparatorRegex,
+    excludeCommandsToOtherBots: Boolean = true,
     scenarioReceiver: CustomBehaviourContextAndTwoTypesReceiver<BC, Unit, TextMessage, Array<String>>
 ): Job = onCommandWithArgs(
     commandRegex = command.toRegex(),
@@ -345,6 +426,7 @@ fun <BC : BehaviourContext> BC.onCommandWithArgs(
     subcontextUpdatesFilter = subcontextUpdatesFilter,
     markerFactory = markerFactory,
     argsSeparator = argsSeparator,
+    excludeCommandsToOtherBots = excludeCommandsToOtherBots,
     scenarioReceiver = scenarioReceiver
 )
 
@@ -360,6 +442,7 @@ fun <BC : BehaviourContext> BC.onCommandWithArgs(
     markerFactory: MarkerFactory<in TextMessage, Any>? = ByChatMessageMarkerFactory,
     additionalSubcontextInitialAction: CustomBehaviourContextAndTwoTypesReceiver<BC, Unit, Update, TextMessage>? = null,
     argsSeparator: Regex = TelegramBotCommandsDefaults.defaultArgsSeparatorRegex,
+    excludeCommandsToOtherBots: Boolean = true,
     scenarioReceiver: CustomBehaviourContextAndTwoTypesReceiver<BC, Unit, TextMessage, Array<String>>
 ): Job = onCommandWithArgs(
     command = botCommand.command,
@@ -367,6 +450,7 @@ fun <BC : BehaviourContext> BC.onCommandWithArgs(
     subcontextUpdatesFilter = subcontextUpdatesFilter,
     markerFactory = markerFactory,
     argsSeparator = argsSeparator,
+    excludeCommandsToOtherBots = excludeCommandsToOtherBots,
     scenarioReceiver = scenarioReceiver
 )
 
@@ -383,6 +467,7 @@ fun <BC : BehaviourContext> BC.onCommandWithNamedArgs(
     additionalSubcontextInitialAction: CustomBehaviourContextAndTwoTypesReceiver<BC, Unit, Update, TextMessage>? = null,
     argsSeparator: Regex = TelegramBotCommandsDefaults.defaultArgsSeparatorRegex,
     nameArgSeparator: Regex = TelegramBotCommandsDefaults.defaultNamesArgsSeparatorRegex,
+    excludeCommandsToOtherBots: Boolean = true,
     scenarioReceiver: CustomBehaviourContextAndTwoTypesReceiver<BC, Unit, TextMessage, List<Pair<String, String>>>
 ) = commandWithNamedArgs(
     commandRegex,
@@ -391,6 +476,7 @@ fun <BC : BehaviourContext> BC.onCommandWithNamedArgs(
     markerFactory = markerFactory,
     argsSeparator = argsSeparator,
     nameArgSeparator = nameArgSeparator,
+    excludeCommandsToOtherBots = excludeCommandsToOtherBots,
     scenarioReceiver = scenarioReceiver,
 )
 
@@ -407,6 +493,7 @@ fun <BC : BehaviourContext> BC.onCommandWithNamedArgs(
     additionalSubcontextInitialAction: CustomBehaviourContextAndTwoTypesReceiver<BC, Unit, Update, TextMessage>? = null,
     argsSeparator: Regex = TelegramBotCommandsDefaults.defaultArgsSeparatorRegex,
     nameArgSeparator: Regex = TelegramBotCommandsDefaults.defaultNamesArgsSeparatorRegex,
+    excludeCommandsToOtherBots: Boolean = true,
     scenarioReceiver: CustomBehaviourContextAndTwoTypesReceiver<BC, Unit, TextMessage, List<Pair<String, String>>>
 ) = onCommandWithNamedArgs(
     command.toRegex(),
@@ -415,6 +502,7 @@ fun <BC : BehaviourContext> BC.onCommandWithNamedArgs(
     markerFactory = markerFactory,
     argsSeparator = argsSeparator,
     nameArgSeparator = nameArgSeparator,
+    excludeCommandsToOtherBots = excludeCommandsToOtherBots,
     scenarioReceiver = scenarioReceiver
 )
 
@@ -431,6 +519,7 @@ fun <BC : BehaviourContext> BC.onCommandWithNamedArgs(
     additionalSubcontextInitialAction: CustomBehaviourContextAndTwoTypesReceiver<BC, Unit, Update, TextMessage>? = null,
     argsSeparator: Regex = TelegramBotCommandsDefaults.defaultArgsSeparatorRegex,
     nameArgSeparator: Regex = TelegramBotCommandsDefaults.defaultNamesArgsSeparatorRegex,
+    excludeCommandsToOtherBots: Boolean = true,
     scenarioReceiver: CustomBehaviourContextAndTwoTypesReceiver<BC, Unit, TextMessage, List<Pair<String, String>>>
 ) = onCommandWithNamedArgs(
     botCommand.command,
@@ -439,5 +528,6 @@ fun <BC : BehaviourContext> BC.onCommandWithNamedArgs(
     markerFactory = markerFactory,
     argsSeparator = argsSeparator,
     nameArgSeparator = nameArgSeparator,
+    excludeCommandsToOtherBots = excludeCommandsToOtherBots,
     scenarioReceiver = scenarioReceiver
 )